+++ src/main/java/kono/ceu/gtconsolidate/common/metatileentities/multi/electric/MetaTileEntityTurboBlastFurnace.java
@@ -58,7 +58,9 @@ import gregicality.multiblocks.common.block.GCYMMetaBlocks;
 import gregicality.multiblocks.common.block.blocks.BlockLargeMultiblockCasing;
 import gregicality.multiblocks.common.block.blocks.BlockUniqueCasing;
 
+import kono.ceu.gtconsolidate.GTConsolidateConfig;
 import kono.ceu.gtconsolidate.api.recipes.GTConsolidateRecipeMaps;
+import kono.ceu.gtconsolidate.api.util.GTConsolidateUtil;
 import kono.ceu.gtconsolidate.api.util.mixinhelper.MultiblockDisplayTextMixinHelper;
 import kono.ceu.gtconsolidate.client.GTConsolidateTextures;
 import kono.ceu.gtconsolidate.common.blocks.BlockCoolantCasing;
@@ -131,7 +133,7 @@ public class MetaTileEntityTurboBlastFurnace extends GCYMRecipeMapMultiblockCont
                         tl.add(TextComponentUtil.setHover(body, hoverString));
                     }
                 });
-        ((MultiblockDisplayTextMixinHelper) builder).addExtendedParallelLine(recipeMapWorkable);
+        GTConsolidateUtil.addExtendedParallelLine(builder, recipeMapWorkable);
         builder.addWorkingStatusLine()
                 .addCustom(tl -> {
                     if (!isActive() && isStructureFormed()) {
@@ -153,8 +155,10 @@ public class MetaTileEntityTurboBlastFurnace extends GCYMRecipeMapMultiblockCont
                         tl.add(TextComponentUtil.setHover(body, hover));
                     }
                 });
-        ((MultiblockDisplayTextMixinHelper) builder).addExtendedProgressLine(recipeMapWorkable);
-        ((MultiblockDisplayTextMixinHelper) builder).addOutputLine(recipeMapWorkable);
+        GTConsolidateUtil.addExtendedProgressLine(builder, recipeMapWorkable);
+        if (GTConsolidateConfig.feature.addOutputLine) {
+            ((MultiblockDisplayTextMixinHelper) builder).addOutputLine(recipeMapWorkable);
+        }
     }
 
     @Override
@@ -416,7 +420,6 @@ public class MetaTileEntityTurboBlastFurnace extends GCYMRecipeMapMultiblockCont
                 }
             }
         }
-        setTemperatureBonus(initialTemperature, blastFurnaceTemperature);
     }
 
     @Override
@@ -460,16 +463,8 @@ public class MetaTileEntityTurboBlastFurnace extends GCYMRecipeMapMultiblockCont
         return blastFurnaceTemperature;
     }
 
-    private void setTemperatureBonus(int baseTemp, int currentTemp) {
-        double ratio = (double) currentTemp / baseTemp;
-
-        // For every doubling, multiply duration by 0.95.
-        double durationBonus = Math.pow(0.9, (int) Math.floor(Math.log(ratio) / Math.log(2)));
-        // For every 5 times, multiply EUt by 0.5
-        double eutBonus = Math.pow(0.5, (int) Math.floor(Math.log(ratio) / Math.log(5)));
-
-        recipeMapWorkable.setSpeedBonus(durationBonus);
-        recipeMapWorkable.setEUDiscount(eutBonus);
+    private int getInitialTemperature() {
+        return initialTemperature;
     }
 
     private int getBonus() {
@@ -511,6 +506,23 @@ public class MetaTileEntityTurboBlastFurnace extends GCYMRecipeMapMultiblockCont
 
         public GigaBlastFurnaceRecipeLogic(RecipeMapMultiblockController metaTileEntity) {
             super(metaTileEntity);
+        }
+
+        @Override
+        protected void modifyOverclockPost(int[] resultOverclock, @NotNull IRecipePropertyStorage storage) {
+            super.modifyOverclockPost(resultOverclock, storage);
+
+            double ratio = (double) getInitialTemperature() / ((IHeatingCoil) metaTileEntity).getCurrentTemperature();
+
+            // For every 5 times, multiply EUt by 0.5
+            double eutBonus = Math.pow(0.5, (int) Math.floor(Math.log(ratio) / Math.log(5)));
+            resultOverclock[0] = (int) (resultOverclock[0] / eutBonus);
+            // For every doubling, multiply duration by 0.9.
+            double durationBonus = Math.pow(0.9, (int) Math.floor(Math.log(ratio) / Math.log(2)));
+            resultOverclock[1] = (int) (resultOverclock[1] / durationBonus);
+
+            resultOverclock[0] = Math.max(1, resultOverclock[0]);
+            resultOverclock[1] = Math.max(1, resultOverclock[1]);
         }
 
         @Override